test: OCD-2593-Verify warnings for duplicate values in criteria detail fields in upload template
