--- src/test/java/gov/healthit/chpl/aqa/RestHeaderDefinitions/HeaderDefinitions.java
@@ -1,8 +1,0 @@
-package gov.healthit.chpl.aqa.RestHeaderDefinitions;
-
-/**
- * Class HeaderDefinitions definition.
- */
-public class HeaderDefinitions {
-
-}
