--- src/test/java/gov/healthit/chpl/aqa/RestHeaderDefinitions/package-info.java
@@ -1,4 +1,0 @@
-/**
- * All REST header definitions.
- */
-package gov.healthit.chpl.aqa.RestHeaderDefinitions;
