+++ src/test/java/gov/healthit/chpl/aqa/asserts/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Assert statements grouped per page for better framework maintenance.
+ */
+/**
+ * @author amore
+ *
+ */
+package gov.healthit.chpl.aqa.asserts;
